keep physics specific code in PhysicsEngine
